moves old implementations of sensors to compat
