+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sensors/compat/camera/__init__.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2022, NVIDIA CORPORATION & AFFILIATES, ETH Zurich, and University of Toronto
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""
+Camera wrapper around USD camera prim to provide an interface that follows the robotics convention.
+"""
+
+from omni.isaac.orbit.sensors.camera.camera_cfg import FisheyeCameraCfg, PinholeCameraCfg
+
+from .camera import Camera, CameraData
+
+__all__ = ["Camera", "CameraData", "PinholeCameraCfg", "FisheyeCameraCfg"]
