+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sensors/compat/height_scanner/height_scanner.py
